@@ -1,13 +1,13 @@
 {
     "private": true,
     "scripts": {
-        "dev": "npm run development",
         "development": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --progress --config=node_modules/laravel-mix/setup/webpack.config.js",
-        "watch": "npm run development -- --watch",
-        "watch-poll": "npm run watch -- --watch-poll",
+        "watch": "npm run dev -- --watch",
+        "watch-poll": "npm run dev -- --watch-poll",
         "hot": "cross-env NODE_ENV=development node_modules/webpack-dev-server/bin/webpack-dev-server.js --inline --hot --config=node_modules/laravel-mix/setup/webpack.config.js",
-        "prod": "npm run production",
-        "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --config=node_modules/laravel-mix/setup/webpack.config.js"
+        "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --config=node_modules/laravel-mix/setup/webpack.config.js",
+        "dev": "vite",
+        "build": "vite build"
     },
     "devDependencies": {
         "@popperjs/core": "^2.11.6",
@@ -16,7 +16,6 @@
         "bootstrap": "^5.2.3",
         "cross-env": "^5.1",
         "jquery": "^3.2",
-        "laravel-mix": "^6.0.49",
         "lodash": "^4.17.5",
         "popper.js": "^1.12",
         "postcss": "^8.4.38",
@@ -24,7 +23,10 @@
         "sass": "^1.56.1",
         "sass-loader": "^7.1.0",
         "tailwindcss": "^3.4.4",
-        "vue": "^2.5.17",
-        "vue-template-compiler": "^2.7.16"
+        "vue": "^2.7.0",
+        "vue-template-compiler": "^2.7.16",
+        "vite": "^3.0.2",
+        "laravel-vite-plugin": "^0.6.0",
+        "@vitejs/plugin-vue2": "^1.1.2"
     }
 }
